# Conflicts:
#	config/application.properties
#	stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/BlocksStoreConfiguration.java
#	stores/blocks/src/main/java/com/bloxbean/cardano/yaci/store/blocks/storage/impl/BlockStorageReaderImpl.java

+++ stores/script/src/main/java/com/bloxbean/cardano/yaci/store/script/domain/Redeemer.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.yaci.store.script.domain;
 
-import com.bloxbean.cardano.client.plutus.spec.ExUnits;
-import com.bloxbean.cardano.client.plutus.spec.RedeemerTag;
+import com.bloxbean.cardano.yaci.core.model.ExUnits;
+import com.bloxbean.cardano.yaci.core.model.RedeemerTag;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -9,8 +9,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.math.BigInteger;
 
 @Getter
 @AllArgsConstructor
@@ -20,7 +18,7 @@ import java.math.BigInteger;
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
 public class Redeemer {
     private RedeemerTag tag;
-    private BigInteger index;
+    private Integer index;
     private String data;
     private ExUnits exUnits;
 }
